improok-social-media/social_media_app/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.37.0
colorama==0.4.6
cryptography==41.0.7
Deprecated==1.2.14
Django==4.2.7
django-celery-results==2.5.1
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-js-asset==2.1.0
django-oauth-toolkit==2.3.0
django-redis==5.4.0
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
jwcrypto==1.5.0
kombu==5.3.5
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
prompt-toolkit==3.0.43
pycparser==2.21
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
